only render boom HTML for app route
