+++ src/apiProxy/apiProxyRoutes.js
@@ -26,35 +26,32 @@ const getApiProxyRoutes = (path, env, apiProxyFn$) => {
 	 * @returns Array query responses, which are in the format defined
 	 *   by `apiAdapter.apiResponseToQueryResponse`
 	 */
-	const handler = getApiProxyRouteHandler(proxyApiRequest$);
-	const plugins = {
-		'electrode-csrf-jwt': {
-			enabled: true,
-		}
-	};
-	const app = {
-		jsonError: true,
-	};
-
-	const apiGetRoute = {
+	const routeBase = {
 		path,
-		handler,
+		handler: getApiProxyRouteHandler(proxyApiRequest$),
+		config: {
+			plugins: {
+				'electrode-csrf-jwt': {
+					enabled: true,
+				}
+			},
+		},
+	};
+	const apiGetRoute = {
+		...routeBase,
 		method: ['GET', 'DELETE', 'PATCH'],
 		config: {
-			app,
-			plugins,
+			...routeBase.config,
 			validate: {
 				query: validApiPayloadSchema
 			},
 		},
 	};
 	const apiPostRoute = {
-		path,
-		handler,
+		...routeBase,
 		method: 'POST',
 		config: {
-			app,
-			plugins,
+			...routeBase.config,
 			validate: {
 				payload: validApiPayloadSchema
 			},
@@ -65,3 +62,4 @@ const getApiProxyRoutes = (path, env, apiProxyFn$) => {
 };
 
 export default getApiProxyRoutes;
+
