+++ src/util/serverUtils.js
@@ -1,6 +1,4 @@
 import https from 'https';
-import React from 'react';
-import ReactDOMServer from 'react-dom/server';
 import Hapi from 'hapi';
 
 import track from './tracking';
@@ -37,24 +35,6 @@ export function configureEnv(config) {
 }
 
 /**
- * This function provides global error handling when there is a 500 error
- */
-export function onPreResponse(request, reply) {
-	const response = request.response;
-	if (!response.isBoom || ((request.route.settings || {}).app || {}).jsonError) {
-		return reply.continue();
-	}
-	const error = response;
-	const { RedBoxError } = require('redbox-react');
-	const errorMarkup = ReactDOMServer.renderToString(
-		React.createElement(RedBoxError, { error })
-	);
-	const errorResponse = reply(`<!DOCTYPE html><html><body>${errorMarkup}</body></html>`);
-	errorResponse.code(error.output.statusCode);
-	return errorResponse;
-}
-
-/**
  * server-starting function
  */
 export function server(routes, connection, plugins, platform_agent, config) {
@@ -64,7 +44,6 @@ export function server(routes, connection, plugins, platform_agent, config) {
 
 	return server.connection(connection)
 		.register(plugins)
-		.then(() => server.ext('onPreResponse', onPreResponse))
 		.then(() => server.auth.strategy('default', 'oauth', true, config))
 		.then(() => server.log(['start'], `${plugins.length} plugins registered, assigning routes...`))
 		.then(() => server.route(routes))
